buy sUSD internally instead of using Depot
The lack of liquidity in the Depot has caused issues in the past.
Here we use a pattern that is found in Exchanger._exchange, for
internally swapping ETH to sUSD.

This approach should be less brittle, as no liquidity is needed for the
swap.

Cheers to @jacko125 for this suggestion.

This commit also permissions the ETHWrapper to call
FeePool internal functions, in order to better report
fees paid using recordFeePaid.