fix: adjust how burn calculation is made
Due to using multiplyDecimalRound in burn() and then divideDecimalRound in _burn(), there would be precision loss. This could mean the contract capacity can't be drained.

This adjusts the order of operations, such that the principal is passed down rather than the full amount, avoiding the loss of precision and thus allowing the full contract to be drained.

A test will be added in a following commit for this.