* test for distributeFees * schedule wrappr fee distribution into feepool period
Jacko and I discussed this change. There were a couple tradeoffs - by distributing fees in sUSD, stakers are always exposed to ETH currency risk. In the previous design, fees were distributed in sUSD at the time of mint/burn, which exposed stakers to a realtime currency risk. We thought about allowing the fee pool to accept sETH as payment, though it would be too much work. So instead, we decided to escrow the fees as sETH for the duration of the fee period, and after the period closes, only then we convert it to sUSD and pay it out.