remove gasPrice overrides from simulate-deploy
this can cause CI-pipeline failure upon release if `gasPrice` is specified with the new EIP-1559 changes. EIP-1559 imposes a minimum gas price, and therefore setting `gasPrice: 0` would cause failure.

I'm not sure if removing the option here was the best to take to verify functionality, please correct if this is not the right approach for this workflow.