fix duplicate fee reporting on WrapperFactory
It is indeed `WrapperFactory`, its submitting a second `recordFeePaid`. now submitting PR. We will also need to reimport fee periods with correct values, or fund the Fee pool the $257k somehow

as for why, there is something wrong with the unit testing framework's handling of the fee address, because this change causes 2 test failures in 2 separate contract tests (one existing in FeePool, and one I just added for completeness) which verify `feesToDistribute` is set correctly, even though as we have now seen on chain this not how it works. as you can see on this PRs run, those tests are now failing even though they check for the correct value.