fix exchange before reclaim risk
`Exchanger.exchange` does not use the post-reclaim amount to calculate the amount of exchange for the
user trade. This leads to the user being left with more of their source token then they should
have had.

this PR resolves this problem by ensuring amounts used for exchange are
calculated for the amount after fee reclamation has occured. It also adds associated tests.